Login and Signup moved to proper component
